Disable the Debug Heap In All Cases
posix_memalign gets patched out which causes DebugHeap::memalign to
crash. This isn't usually a problem since the DebugHeap isn't used.
But, it can be activiated by loading ASAN or TSAN libraries into the
process. This makes it impossible to use these types of tools to
debug your own native code running along-side JSC. This patch
modifies the method that makes the decision as to use the DebugHeap
so that it always returns false.