Fix DebugHeap Crash
Previously JSC's debug heap would crash when activated by using
tools like ASAN. This was due to a call to posix_memalign() being
patched out during the build process which caused
DebugHeap::memalign() to unconditionally call BCRASH(). The call
to posix_memalign() was originally patched out because it is not
offered in Android API <= 16 and jsc-android targets API >= 16.
This change replaces the original patch with one that adds an
implementation of posix_memalign() borrowed from the Android
Support Library source.